JS Animation
Began the JS setup for wave effect on labels element. By creating direct link to to each letter. Once I was able span each letter went back to css to style labels. Very nice transition with transform to give it a bounce upward movement. Once that was done added into the span a delay for each letter to give it the wavy transition.