Javascript for Big Cup
Called the function for the big cup in 2 parts. One at the beginning were the application load, and one where the highlights function is at. Now a function is created for the big cup, adding syntax for CSS selectors to be selected according to the length (or quantity of cups selected).  Added if block for to hide percentage when empty, then added else block to make percentage visible  with amount of full cups highlighted and to show percentage that is full. Added if block for when big cup is full it will hide the remaining Liters. Added else block for the remaining liter to be visible when the big cp is not full, and added inner text for the span for liters.