Javascript for Movie App
Registered to receive an API key from 3rd party movie website. Created a constant to access to popular movie url. Created a constant for access to the image path and a constant for the search url.

Added a request (async function)  to get movies by taking in a url, creating a variable of response to a promise (fetch)  passing in a url. Created another variable for data to  await a response from json to transmit the data objects.

Create a call to get initial movies passing in the initial API URL.

Create a call for the variable of form, to listen for a submit and to not submit to the main page, then create another variable to search  with a value within the form variable, make sure to add  a variable for search  to the call list on top section of the script.

Then check to see if search term exists and if search term doesn't equal to anything. Then call the variable to get movies  by passing search API and then add the search term.

Then clear the search value.  If a submit is done and nothing is written inside the form then window will reload to the initial url or API URL.