Javascript Pt.2 Adding Movies to the Dom
Having created a a fetch and search functions, now we must create functions to actually get the the movies and post them on the web page. We start of within async function by pluggin in  an array. Then outside the async function we create a function for that array to take in the data. We clear out the main webpage to be able to plug in the movies. And then make sure to plug it in to be accessed globally in the code by creating array at the beginning of the code as a constant.
For the data thats fetched we will loop thru it, deconstruct to get the title, poster, vote and overview.
The we will construct the div into the dom to present the fetched data that way. We will be plugging in one of the divs on the HTML and delete the 3 divs we created adding it to the inner HTML. On the pasted div, we will add  expression syntax for the poster, title, vote, and overview.
Create function for vote, this will give a range for each color to be presented .