HTML and Javascript
Added font awesome links, with a main container. Within main container added 5 container that include background images that will appear when the slide active is toggled with the buttons (one for the right and one for the left).

Added styling to the body and image found within the body, slide container,  containers within the slide container and buttons.