HTML and CSS
Added a button with toggling to change from light mode to dark mode.Added all classes needed to style the clock, clock needles with every needle having its own class and the center point of the clock as well. Added more classes for the time, date and within the date container text for styling as well.

Added root to access colors globally. Styled html for hen the dark mode is toggled, as well as the styling for the dark mode. Styling to the body and classes within the body.

I would like to not in the class of clock styling background color originally was there to have a visual on positioning of the clock once everything for the clock was set up, we eliminated due to it serving its purpose.