Javascript for Theme Clock
Added all elements needed to select globally. Continued with an event listener for the toggle to switch between light and dark mode.

Added a function for the time that will be set, by adding all variables needed. Making sure to call the function globally (setTime()).

Added the needles for hour, minute, and seconds as elements to have them reflect the actual time, I used the same code from the CSS but changed the rotate by making it dynamic thru calling a scale function. Once that function is called I added the necessary parameters for hours, minutes and seconds.

Once thats done the time will show but it wont continuously show the time. In order to do that we must call it globally by adding set time function with the time handler or a setInterval.

Added time and date to display real time and date HTML.  We must also add  month and days arrays.