HTML and CSS
At first we created a button that has text and a span with a class and style property. Once this we are able to get a visual in our CSS of the button we will eliminate the span and add it to Javascript. I also added a class to the button due to seeing the global view of a website. Even though this is one button its better practice to give the button a class for when actually designing a website you will have many buttons.

Added styling to the button to the class of circle worked with scaling within the transform styling which we had to access keyframes and then add the same transform style but changing the value for the scale and adding an opacity to make the circle dissapear.