Javascript
We begin by documenting to all buttons with the ripple class. We create the method for each  button and add an event listener then we add variables for accessing the click on the x and y axis within the button. We continue adding variables now to measure/pinpoint within the button the click.

Here is where we construct the span that was eliminated on the HTML. We add the variable for circle found in CSS that documents a created element of span and then add class for circle and then access that class to style it for variables that access the measuring of the click.

Important note, a problem occurs for ever click a new span for the circle is created and we don't want it to be created but to disappear. In order to have the DOM eliminate that span we set a timeout to the circle so it can be removed.