Drawing App HTML and CSS
Added empty canvas element, since we will be drawing, a container for the toolbox with 3 button and input type of color and a span with text.

Added styling for the body and canvas elements. As well as the toolbox class, the direct children for the toolbox class and the last child of the toolbox class.