Javascript for the Drawing App
Added the necessary variables with their method first one is to call the canvas element, and the second one is to get access to the context of that element.

The created a function (drawCircle) for positioning on the x and y axis, a function (drawLine) for drawing. Made sure to create variables globally for both of those functions.
Then added event listeners for movement of the mouse, if its pressed (mousedown) downward and when I let go (mouseup). Added another variable (isPressed) globally and set to false. This means when I do hold down on the mouse it will be set on that function to true also set to let the DOM know here in that context the mouse is being held down. Once I let go of the mouse it will switch back the variable (isPressed) to false and not define where in the context we stop holding the mouse.
Added and if statement for when the mouse is held down, and then called the draw circle and draw line function so that a combination of a circle and line are drawn when the mouse is held down. Because the line width size is measured different than arc size we must multiply the  line width times 2.