Drawing App JS Update
Changed the value of the size variable.

Added a couple more variables for the next steps. We will be giving functionality to the increase button, decrease button, color  and clear elements.
Started with adding an event listener to the color element to be able to change the color value.
The continued with the increase button adding and event listener as well, so that when the button is clicked it will increase the size by 5, and will check if it is greater than 50 but we make sure to set it size to 50 so it will not be greater than 50. We must make sure this value is updated on the screen so we add the updatesizeonscreen function inside the event listener for the increase button.
We do the exact same code for the decrease button but we change the size to decrease (-=) and to check the size to less than 5 and set the size to be 5.

Add function for the update we need for increeases and decreases.

Lastly add an event listener for clear element to clear out x, y and the whole canvas area.