Kinetic CSS Loader
For the HTML I added a container with a class of kinetic.

In CSS using that one class of kinetic we do all the styling for the loader. I started by setting the universal styling. We then move on to the body styling. then move on to the styling for the class.
Then we split the class to have 2 shapes (after and a before), add it styling.
Then we focus on rotating only 1 of the shapes (before) to a starting point that will give it the kinetic effect. Then we add animation to both shapes.
We access keyframes and add the styling needed for the rotations to be done.