Content Place Holder
Added containers with class for the card (card), inside the card container will be a header with animation (card-header, animated-bg) and an image. Another container with what the card will contain (card content), inside this container will be heading with a class, id and template writing, a paragraph with a class, id, template writing and 3 span elements each with their class and a non breaking space (&nbsp)(these span are for the loading animation).
Another container with the author details, inside this container will be a container with an image and another container for the authors info which will have its class and id with strong text for the name; and small text with its class and id for the date.

CSS
Added styling for classes. For animation I added linear gradient with a movement to the right with 4 different color for percentage of the gradient. Added a background size for and animation with name.
Added keyframes for the animation name (bgPos) created.