Revised HTML Javascript Integration
Everything that was omitted was part of what we had added for the info we wanted to show up in the card. All the images, text, name and date. We replaced them with a non breaking space (&nbsp).

Made sure to create all variables we need to access globally.
Then continued by adding a function to access the data we replaced on the HTML. Added inside the function all Id's I need to access. Lastly added 2 forEach function to remove the content found in the named classes.